requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
sqlalchemy>=2.0.29
alembic>=1.13.1
psycopg2-binary>=2.9.9
pydantic>=2.7.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.2
google-generativeai>=0.5.0
chardet>=5.2.0,<6.0.0
google-genai>=1.0.0
truststore>=0.9.2
certifi>=2024.8.30
email-validator>=2.1.0.post1
